Applied ML/AI Agents/data-onboarding/pyproject.toml
[project]
name = "data-onboarding"
version = "0.1.0"
description = "A multi-agent ADK system that automates data onboarding into BigQuery from URLs or GCS paths."
readme = "readme.md"
requires-python = ">=3.13.3"
dependencies = [
    "google-adk>=1.14.0",
    "google-genai>=1.64.0",
    "google-cloud-aiplatform[agent_engines,adk]>=1.88.0",
    "google-cloud-bigquery>=3.31.0",
    "google-cloud-storage>=2.19.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.12.5",
    "beautifulsoup4>=4.13.0",
    "httpx>=0.28.0",
    "pandas>=2.3.3",
    "openpyxl>=3.1.0",
    "pdfplumber>=0.11.0",
    "google-cloud-datacatalog-lineage>=0.3.0",
    "google-cloud-geminidataanalytics>=0.3.0",
    "vl-convert-python>=1.8.0",
    "altair>=5.5.0",
]

[project.optional-dependencies]
dev = [
    "db-dtypes>=1.5.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"
extend-exclude = ["examples", "output", "deploy", "scripts"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP", "SIM"]
ignore = ["E501", "B008", "ARG001"]

[tool.pytest.ini_options]
testpaths = [
    "agent_orchestrator/tests",
    "agent_acquire/tests",
    "agent_discover/tests",
    "agent_understand/tests",
    "agent_design/tests",
    "agent_implement/tests",
    "agent_validate/tests",
    "agent_context/tests",
    "agent_convo/tests",
]
asyncio_mode = "auto"
addopts = "--cov=agent_orchestrator --cov=agent_acquire --cov=agent_discover --cov=agent_understand --cov=agent_design --cov=agent_implement --cov=agent_validate --cov=agent_context --cov=agent_convo --cov-report=term-missing"

[tool.coverage.run]
source = [
    "agent_orchestrator",
    "agent_acquire",
    "agent_discover",
    "agent_understand",
    "agent_design",
    "agent_implement",
    "agent_validate",
    "agent_context",
    "agent_convo",
]

[tool.coverage.report]
show_missing = true
skip_empty = true
